Add support for POSTBOX_READY as a response
This commit will allow the postbox flow to continue working when a
callback is returned with a `POSTBOX_READY` on top of the expected `SUCCESS`.

This is needed because Android digi.me application currently returns
`POSTBOX_READY` on successful postbox creation. That will be changed in
future releases of the android application.

Until then, this example application will need to support both success
flags.